requirements.txt
# Core ML and Data Processing
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1

# Deep Learning and NLP
torch==2.0.1
transformers==4.31.0
tokenizers==0.13.3
datasets==2.14.0

# Text Processing
spacy==3.6.0
textstat==0.7.3
python-docx==0.8.11
PyPDF2==3.0.1
pymupdf==1.23.7

# Web Framework and API
flask==2.3.2
uvicorn[standard]==0.23.2
pydantic==1.10.8
gunicorn==21.2.0

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.65.0

# Development and Testing
pytest==7.4.0
pytest-asyncio==0.21.1
black==23.7.0
flake8==6.0.0

# Jupyter and Visualization
jupyter==1.0.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Model Persistence
joblib==1.3.2
pickle-mixin==1.0.2

# OCR and Image Processing (Bonus Features)
pytesseract==0.3.10
Pillow==10.0.0
opencv-python==4.8.0.76
pdf2image==1.16.3
pdfplumber==0.9.0

# Monitoring and Logging
loguru==0.7.0
prometheus-client==0.17.1

# Additional Dependencies
evaluate==0.4.0
setuptools>=65.0.0

# Additional ML and Data Science
scikit-learn-intelex==2023.1.1
nltk==3.8.1
wordcloud==1.9.2

# Additional Utilities
requests>=2.31.0
urllib3>=2.0.0

# Production and Deployment
waitress>=2.1.2
sentry-sdk[flask]>=1.28.0
redis>=4.5.0

# Additional Development Tools
ipython>=8.0.0
notebook>=6.5.0 